requirements.txt
requests==2.31.0
pyTelegramBotAPI==4.14.0
python-telegram-bot==13.15
python-dotenv==1.0.0
tabulate==0.9.0
fastf1
pandas
pika